+++ main.lua
@@ -1,103 +1,27 @@
 require('mobdebug').start();              -- enable debugging checkpoints
 StartDebug();                             -- enable debugging
-local mod = RegisterMod("AI Final", 1);   -- register mod in game
+mod = RegisterMod("AI Final", 1);   -- register mod in game
 
-PriorityQueue = require("priority_queue") -- import priority queues
-
--- room search:
+---------------------------
+-------- TODO LIST --------
+---------------------------
+--[[
+aStar Room Search :
 -- add cost for spider webs, chests, etc.
--- Smoothing algorithm for aStarRoomSearch needs to be changed
-  -- isDirectPath doesn't take into account entities
 -- also we need to let isaac know that he can break POOP and FIRE tiles if he wants to get through them
 -- also we need to special case the on and off spikes, since isaac needs to go over them in some cases when they are off
-
 -- TODO nextValidGridIndices needs to be updated to deal with diagonal movement blocking cases
+-- TODO increase cost of moving diagonally in aStarRoomSearch
+]]
 
--- TODO add support for other types
-function num(value)
-  local value_type = type(value)
-  if value_type == "boolean" then
-    return value and 1 or 0
-  end
-  return nil
+-------------------------
+-------- IMPORTS --------
+-------------------------
+
+-- Write str to the Isaac Console
+function CPrint(str) 
+  Isaac.ConsoleOutput(str.."\n")
 end
-
--- swaps key and value pairs
-function makeReverseTable(someTable)
-  local newTable = {}
-  for index, val in pairs(someTable) do
-    newTable[val] = index
-  end
-  return newTable
-end
-
--- converts a decimal number to a list of booleans representing the binary conversion
-function toBitBools(num)
-    -- returns a table of bits, least significant first.
-    local t={} -- will contain the bits
-    local i=1
-    while num>0 do
-        rest=math.fmod(num,2)
-        t[i]=(rest == 1)
-        i = i + 1
-        num=(num-rest)/2
-    end
-    return t
-end
-
-function isEntityBossDying(entity)
-  return isEntityFlagTrue(entity, 20)
-end
-
--- returns if the given entity has the EntityFlag true
-function isEntityFlagTrue(entity, flagNum)
-  local flags = toBitBools(entity:GetEntityFlags())
-  return flags[flagNum]
-end
-
--- is any entity in this room a boss dying?
-function isBossDying()
-  for index, entity in pairs(getAllRoomEntities()) do
-    if isEntityBossDying(entity) then
-      return true
-    end
-  end
-  return false
-end
-
------------------------------------
--------- GLOBAL VARIABLES ---------
------------------------------------
--- reverse tables for enums
-GridEntityEnumReverse = makeReverseTable(GridEntityType)
-DoorSlotEnumReverse = makeReverseTable(DoorSlot)
-RoomTypeEnumReverse = makeReverseTable(RoomType)
-
--- Level search. Instantiated by gameStart
-local dfsIterator = nil
-local directions = nil
-local goalTest = nil
-local timer = 0
-local moveLeftAndRightEvery = 100
-
-local AgentType = { MoveLeftAndRight = 0, SnakeAgent = 1,  DumbPointAndClick = 2, SmartPointAndClick = 3, SmartBoiAgent = 4 }
-local agentType = AgentType.SmartBoiAgent
-local agentTypeString = "SmartBoiAgent"
-
-local isaacMessage = ""
-local isaacMessageTimer = 0
-local isaacMessageTimerInitValue = 0
-
-local pointAndClickPos = nil
-local pointAndClickThreshold = 20
-
- -- position to go to returned from DFS level search
-local levelSearchDoorPosition = nil
-local shouldRunLevelSearch = false
-
------------------------------------
--------- UTILITY FUNCTIONS --------
------------------------------------
 
 function import(filename)
   local _, err = pcall(require, filename)
@@ -110,261 +34,61 @@ function import(filename)
   end
 end
 
--- Write str to the Isaac Console
-function CPrint(str) 
-  Isaac.ConsoleOutput(str.."\n")
-end
+-- data structures
+import("stack")
+import("priority_queue")
 
--- Write str to file specified by config (or hardcoded)
-function log(str)
-  -- TODO change this to write to a log file.
-  CPrint(str)
-end
+-- utility functions
+import("lua_utils")
+import("print_utils")
+import("isaac_utils")
+import("isaac_types")
+import("pedestal_items")
 
--- Utility method for testing
-function equal(expected, result)
-  -- TODO make this smart?
-  return expected == result
-end
+-- callback hooks
+import("on_game_start")
+import("on_level_start")
+import("on_room_start")
+import("on_damage") -- doesn't do anything yet, might want to use this for scoring the agent
+import("on_input_request")
+import("on_step")
 
--- Check membership of val in list
-function contains(list, val)
-    for index, value in ipairs(list) do
-        if value == val then
-            return true
-        end
-    end
+-- AI implementations and drivers
+import("agents")
+import("room_search")
+import("navigator")
 
-    return false
-end
+-----------------------------------
+-------- GLOBAL VARIABLES ---------
+-----------------------------------
 
--- Basic functional mapping
-function map(func, array)
-  local new_array = {}
-  for i,v in ipairs(array) do
-    new_array[i] = func(v)
-  end
-  return new_array
-end
+-- Level search. Instantiated by gameStart
+local directions = nil
+local goalTest = nil
 
--- Basic functional mapping
-function filter(filter_func, array)
-  local new_array = {}
-  for i,v in ipairs(array) do
-    if filter_func(v) then
-      table.insert(new_array, v)
-    end
-  end
-  return new_array
-end
-
--- return a new array equal to old plus one new element elt
-function append(array, elt)
-  local newArr = {}
-  for k, v in ipairs(array) do
-    newArr[k] = v
-  end
-  newArr[#array+1] = elt
-  return newArr
-end
-
--- return a new array equal to old plus one new element elt
-function appendAtIndex(array, elt, index)
-  local newArr = {}
-  for k, v in ipairs(array) do
-    if k < index then
-      newArr[k] = v
-    elseif k == index then
-      newArr[k] = elt
-      newArr[k+1] = v
-    else
-      newArr[k+1] = v
-    end
-  end
-  return newArr
-end
-
-
--- retrieves the player userdata
-function getPlayer()
-  return Isaac.GetPlayer(0)
-end
-
--- get the position of the player in-game
--- these coordinates are used for doing stuff in-game
-function getPlayerPosition()
-  return getPlayer().Position
-end
-
--- get the position of the player on the screen
--- these coordinates are used for drawing to the screen
-function getPlayerScreenPosition()
-  return Isaac.WorldToScreen(getPlayerPosition())
-end
-
-function getScreenPosition(pos)
-  return Isaac.WorldToScreen(pos)
-end
-  
 --------------------------------
 -------- PROJECT HEADER --------
 --------------------------------
-Isaac.ConsoleOutput("----------------------------------\n")
-Isaac.ConsoleOutput("--- CS4100 Project Initialized ---\n")
-Isaac.ConsoleOutput("----------------------------------\n")
-Isaac.ConsoleOutput("\n")
+CPrint("----------------------------------")
+CPrint("--- CS4100 Project Initialized ---")
+CPrint("----------------------------------")
+CPrint("")
 
 modEnabled = true
-
-import("levelSearch")
-import("priority_queue")
-
---------------------------------
--------- DEBUG COMMANDS --------
---------------------------------
--- Define debug variables
-local makeIsaacInvincible = true
-local killAllEnemiesOnRoomStart = true
-
-function onGameStarted()
-  Isaac.ConsoleOutput("### New Game Started ###\n")
-  
-  -- Isaac takes damage but his health never decreases
-  if makeIsaacInvincible then
-    Isaac.ExecuteCommand("debug 3")
-  end
-  
-  -- Make all enemies in room die when entering the room
-  if killAllEnemiesOnRoomStart then
-    Isaac.ExecuteCommand("debug 10")
-  end
-  
-  -- print variables to console
-  Isaac.ConsoleOutput(string.format("makeIsaacInvincible = %s\n", tostring(makeIsaacInvincible)))
-  Isaac.ConsoleOutput(string.format("killAllEnemiesOnRoomStart = %s\n", tostring(killAllEnemiesOnRoomStart)))
-end
-
--- bind the MC_POST_GAME_STARTED callback to onGameStarted
--- this event is called when a new game is started or a game is loaded from a save state
-mod:AddCallback(ModCallbacks.MC_POST_GAME_STARTED, onGameStarted)
-
-------------------------------------
--------- PROGRAMMATIC INPUT --------
-------------------------------------
-shootDirection = ButtonAction.ACTION_SHOOTUP
-moveDirectionX = ButtonAction.ACTION_LEFT
-moveDirectionY = nil
-
-function onInputRequest(_, entity, inputHook, buttonAction)
-  if modEnabled then
-    if entity ~= nil then
-      if inputHook == InputHook.GET_ACTION_VALUE then
-        if buttonAction == moveDirectionX or buttonAction == moveDirectionY then
-          return 1.0
-        end
-        if buttonAction == shootDirection then
-          return 1.0
-        end
-        return nil
-      elseif inputHook == InputHook.IS_ACTION_PRESSED then
-        if buttonAction == shootDirection then
-          return true
-        end
-      elseif inputHook == InputHook.IS_ACTION_TRIGGERED then
-        -- do something here?
-      end
-      return nil
-    end
-  end
-end
-
--- bind the MC_INPUT_ACTION callback to onInputRequest
-mod:AddCallback(ModCallbacks.MC_INPUT_ACTION, onInputRequest)
-
-function getCurrentRoom()
-  return Game():GetLevel():GetCurrentRoomIndex()
-end
-
-function isBossRoom()
-  return Game():GetRoom():GetType() == RoomType.ROOM_BOSS
-end
-
-function getAllDoors()
-  local room = Game():GetRoom()
-  doors = {}
-  for slot, idx in pairs(DoorSlot) do
-    door = room:GetDoor(idx)
-    if door then
-      table.insert(doors, door)
-    end
-  end
-  return doors
-end
-
-function isDoorUnlocked(door)
-  if not door then
-    return false
-  else
-    return not door:IsLocked()
-  end
-end
-
-function isDoorSecret(door)
-  if not door then
-    return false
-  else
-    return (door.TargetRoomType == RoomType.ROOM_SECRET or door.TargetRoomType == RoomType.ROOM_SUPERSECRET) and door:CanBlowOpen()
-  end
-end
-
--- good doors are unlocked doors that lead to another normal room, the boss, or an item room
-function isGoodDoor(door)
-  return isDoorUnlocked(door) and not isDoorSecret(door) and
-    ((door.TargetRoomType == RoomType.ROOM_DEFAULT) or
-    (door.TargetRoomType == RoomType.ROOM_BOSS) or 
-    (door.TargetRoomType == RoomType.ROOM_TREASURE))
-end
-
-
-
-function makeDoorPair(door)
-  return {door, DoorSlotEnumReverse[door.Slot]}
-end
-
-function getGoodDoors()
-  return filter(isGoodDoor, getAllDoors())
-end
-
-function getTargetRoomIndex(door)
-  return door.TargetRoomIndex
-end
 
 visitedRooms = {}
 roomStack = Stack:new()
 initialRoom = nil -- a triple of roomIndex, door, and path
 currentPath = Stack:new()
 
-function setInitialLevelSearchParams()
-  visitedRooms = {}
-  roomStack = Stack:new()
-  initialRoom = {getCurrentRoom(), {}}
-  roomStack:push(initialRoom)
-end
-
-
-
 function dfsToNextRoom()
   while not roomStack:isEmpty() do
     local topOfStack = roomStack:pop()
     local currentRoom = topOfStack[1]
     local pathToRoom = topOfStack[2]
-    -- if we are in the boss room we want to navigate to the trap door to advance to the next level
-    if isBossRoom() and getTrapDoor() ~= nil then
-      return getTrapDoor().Position
-    end
     if (not visitedRooms[currentRoom]) then
       visitedRooms[currentRoom] = true
-      for indexInList, nextDoor in pairs(getGoodDoors()) do
+      for _, nextDoor in pairs(getGoodDoors()) do
         local nextRoomIndex = getTargetRoomIndex(nextDoor)
         if (not visitedRooms[nextRoomIndex]) then
           roomStack:push({nextRoomIndex, append(pathToRoom, nextRoomIndex)})
@@ -377,1459 +101,3 @@ function dfsToNextRoom()
     return getDoorTo(roomStack:peek()[1]).Position
   end
 end
-
-function getDoorTo(roomIndex)
-  local allDoors = getAllDoors()
-  for listIndex, door in pairs(allDoors) do
-    if getTargetRoomIndex(door) == roomIndex then return door end
-  end
-  return nil
-end
-
-visitedRooms2 = {}
-
-function getNextUnvisitedDoor()
-  local goodDoors = getGoodDoors()
-  for index, door in pairs(goodDoors) do
-    local roomIndex = getTargetRoomIndex(door)
-    if not visitedRooms2[roomIndex] then
-      return door
-    end
-  end
-  -- else return a random door
-  return goodDoors[math.random(#goodDoors)]
-end
-
-function getClosestFromIndices(gridIndexList)
-  local closestDist = manhattanDist(getGridIndex(getPlayerPosition()), gridIndexList[1])
-  local closestIndex = gridIndexList[1]
-  for indexInList, gridIndex in pairs(gridIndexList) do
-    local dist = manhattanDist(getGridIndex(getPlayerPosition()), gridIndex)
-    if dist < closestDist then
-      closestDist = dist
-      closestIndex = gridIndex
-    end
-  end
-  return closestIndex
-end
-
---------------------------------------------
--- THIS IS WHERE WE UPDATE OUR DIRECTIONS --
---------------------------------------------
-function updateWhereToGoNext()
-  -- if we haven't tried to find the path yet
-  if directions == nil then
-    -- if there are enemies in the room fight them
-    if (not noEnemies()) then
-      return
-    end
-     
-    local pressurePlates = getUnpressedPressurePlates()
-    if (#pressurePlates > 0) then
-      directions = getDirectionsTo(getGridPos(getClosestFromIndices(pressurePlates)))
-      directionIndex = 1
-      
-      -- we have reached our button when there is one less button
-      goalTest = function () return #getUnpressedPressurePlates() == #pressurePlates - 1 end
-      return
-    end
-    
-    -- if there are pedestal items in the room, get those first
-    local pedestalItems = getPassivePedestalItems()
-    if (#pedestalItems > 0) then
-      directions = getDirectionsTo(pedestalItems[1].Position)
-      directionIndex = 1
-      goalTest = function () return #getPassivePedestalItems() == 0 end
-      return
-    end
-    
-    -- if there are no enemies then advance to the next room
-    if (not isBossRoom() and noEnemies()) then
-      directions = getDirectionsTo(getNextUnvisitedDoor().Position)
-      directionIndex = 1
-      goalTest = function () return false end
-      return
-    end
-    
-    -- if there are normal items in the room, get them next
-    
-    -- if there is a trapdoor to the next floor, go there next
-    if (getTrapDoor()) then
-      directions = getDirectionsTo(getTrapDoor().Position)
-      directionIndex = 1
-      goalTest = function () return false end
-      return
-    end
-  end
-end
-
---------------------------
---- ON KILL HOOK ---
---------------------------
-function onKill()
-  
-end
-
--- this event is triggered every kill, which is why we are using it to check the last kill
-mod:AddCallback(ModCallbacks.MC_POST_ENTITY_KILL, onKill)
-
--- called whenever you enter a room
-function onRoomStart()
-  pointAndClickPos = nil
-  directions = nil
-  goalTest = nil
-  directionIndex = 1
-  visitedRooms2[getCurrentRoom()] = true
-end
-
--- bind the MC_POST_NEW_ROOM callback
--- this event is triggered every time you enter a room
-mod:AddCallback(ModCallbacks.MC_POST_NEW_ROOM, onRoomStart)
-
-
-function onLevelStart()
-  visitedRooms2 = {}
-end
-mod:AddCallback(ModCallbacks.MC_POST_NEW_LEVEL, onLevelStart)
- 
--------------------------------
--------- DAMAGE EVENTS --------
--------------------------------
-function onPlayerDamage(_,entity,_,_,source)
-  
-end
-
--- bind the MC_ENTITY_TAKE_DMG callback for the Player to onPlayerDamage
-mod:AddCallback(ModCallbacks.MC_ENTITY_TAKE_DMG, onPlayerDamage, EntityType.ENTITY_PLAYER)
-
----------------------------------
--------- FREQUENT CHECKS --------
----------------------------------
-
--- sets the message to print under the player sprite for the given number of frames (duration)
-function setIsaacMessage(message, duration)
-  isaacMessage = message
-  isaacMessageTimer = duration
-  isaacMessageTimerInitValue = duration -- used to fade the text out
-end
-
--- prints the isaacMessage under the player sprite for isaacMessageTimer frames
-function printIsaacMessage()
-  if isaacMessageTimer > 0 then
-    local opacity = isaacMessageTimer / isaacMessageTimerInitValue
-    local screenPos = getPlayerScreenPosition()
-    Isaac.RenderText(isaacMessage, screenPos.X - string.len(isaacMessage) * 3, screenPos.Y, 1, 1, 1, opacity)
-    isaacMessageTimer = isaacMessageTimer - 1
-  end
-end
-
--- returns whether there is an unobstructed line of movement between
--- the given two grid indices
-function isDirectPath(pos1, pos2)
-  return Game():GetRoom():CheckLine(getGridPos(pos1), getGridPos(pos2), 0, 0, true, true)
-end
-
-function getPlayerGridIndex()
-   return getGridIndex(getPlayerPosition())
-end
-
-function getGridIndex(pos)
-  return Game():GetRoom():GetClampedGridIndex(pos)
-end
-
-function getGridPos(index)
-   return Game():GetRoom():GetGridPosition(index)
-end
-
-function manhattanDist(gridIndex1, gridIndex2)
-  local xPos1 = modulo(gridIndex1, getRoomWidth())
-  local yPos1 = math.floor(gridIndex1 / getRoomWidth())
-  local xPos2 = modulo(gridIndex2, getRoomWidth())
-  local yPos2 = math.floor(gridIndex2 / getRoomWidth())
-  return math.abs(xPos1 - xPos2) + math.abs(yPos1 - yPos2)
-end
-
-function modulo(a, b)
-  return (a - math.floor(a/b)*b)
-end
-
-function getRoomWidth()
-  return Game():GetRoom():GetGridWidth()
-end
-
-function getRoomHeight()
-  return Game():GetRoom():GetGridHeight()
-end
-
-function nextValidGridIndices(gridIndex, goalIndex)
-  local nextIndices = {}
-  local adjacentIndices = getAllAdjacentGridIndices(gridIndex)
-  for _, successorIndex in pairs(adjacentIndices) do
-    if not isGridIndexBlocked(successorIndex)
-    or (manhattanDist(gridIndex, goalIndex) == 1 and successorIndex == goalIndex) then
-      nextIndices = append(nextIndices, successorIndex)
-    end
-  end
-  return nextIndices
-end
-
-function getAllAdjacentGridIndices(gridIndex)
-  local roomWidth = getRoomWidth()
-  local roomHeight = getRoomHeight()
-  local nextIndices = {}
-  nextIndices[1] = gridIndex - roomWidth     -- UP
-  nextIndices[2] = gridIndex + roomWidth     -- DOWN
-  nextIndices[3] = gridIndex - 1             -- LEFT
-  nextIndices[4] = gridIndex + 1             -- RIGHT
-  nextIndices[5] = gridIndex - roomWidth - 1 -- TOPLEFT
-  nextIndices[6] = gridIndex - roomWidth + 1 -- TOPRIGHT
-  nextIndices[7] = gridIndex + roomWidth - 1 -- BOTTOMLEFT
-  nextIndices[8] = gridIndex + roomWidth + 1 -- BOTTOMRIGHT
-  return nextIndices
-end
-
--- get the indices of the four diagonals from the given grid index
-function getCornerGridIndices(gridIndex)
-  local roomWidth = getRoomWidth()
-  local roomHeight = getRoomHeight()
-  local nextIndices = {}
-  nextIndices[1] = gridIndex - roomWidth - 1 -- TOPLEFT
-  nextIndices[2] = gridIndex - roomWidth + 1 -- TOPRIGHT
-  nextIndices[3] = gridIndex + roomWidth - 1 -- BOTTOMLEFT
-  nextIndices[4] = gridIndex + roomWidth + 1 -- BOTTOMRIGHT
-  return nextIndices
-end
-
--- verify how this works with all entities
-function isGridIndexBlocked(gridIndex)
-  -- restrict out of bounds indices
-  if (gridIndex > Game():GetRoom():GetGridSize() or gridIndex < 0) then return true end
-  
-  local gridEntity = Game():GetRoom():GetGridEntity(gridIndex)
-  if (gridEntity ~= nil) then
-    local t = gridEntity:GetType()
-    return not (t == nil 
-      or t == 0 
-      or t == GridEntityType.GRID_DECORATION
-      or t == GridEntityType.GRID_SPIKES_ONOFF
-      or t == GridEntityType.GRID_DOOR
-      or t == GridEntityType.GRID_TRAPDOOR
-      or t == GridEntityType.GRID_SPIDERWEB
-      or t == GridEntityType.GRID_PRESSURE_PLATE
-      or t == GridEntityType.GRID_POOP -- if it is destroyed
-      or t == GridEntityType.GRID_ROCK and gridEntity:GetSaveState().State == 2 -- if it is destroyed
-      or t == GridEntityType.GRID_TNT -- if it is destroyed
-      )
-  else
-    -- check if there are any blocking entities in our way
-    local blockingEntities = filter(entityBlocksMovement, getAllRoomEntities())
-    for index, entity in pairs(blockingEntities) do
-      if getGridIndex(entity.Position) == gridIndex then
-        return true
-      end
-    end
-    return false
-  end
-end
-
--- returns the first trapdoor in the room
-function getTrapDoor()
-  local currRoom = Game():GetRoom()
-  local i = 1
-  while i < Game():GetRoom():GetGridSize() do
-    local gridEntity = currRoom:GetGridEntity(i)
-    if gridEntity ~= nil and gridEntity:GetType() == GridEntityType.GRID_TRAPDOOR then -- if we have found a TrapDoor
-      return gridEntity
-    end
-    i = i + 1
-  end
-end
-
--- returns a string representing the enum type of the grid entity at the given index
--- 1 --> GRID_DECORATION
--- 2 --> GRID_ROCK
--- 3 --> GRID_ROCKB
--- 4 --> GRID_ROCKT
--- 5 --> GRID_ROCK_BOMB
--- 6 --> GRID_ROCK_ALT
--- 7 --> GRID_PIT
--- 8 --> GRID_SPIKES
--- 9 --> GRID_SPIKES_ONOFF
--- 10 --> GRID_SPIDERWEB
--- 11 --> GRID_LOCK
--- 12 --> GRID_TNT
--- 13 --> GRID_FIREPLACE
--- 14 --> GRID_POOP
--- 15 --> GRID_WALL
--- 16 --> GRID_DOOR
--- 17 --> GRID_TRAPDOOR
--- 18 --> GRID_STAIRS
--- 19 --> GRID_GRAVITY
--- 20 --> GRID_PRESSURE_PLATE
--- 21 --> GRID_STATUE
--- 22 --> GRID_ROCK_SS
-function getGridType(gridIndex)
-  local gridEntity = Game():GetRoom():GetGridEntity(gridIndex)
-  if gridEntity ~= nil then
-    return GridEntityEnumReverse[gridEntity:GetType()]
-  end
-  return ""
-end
-
--- get all buttons / pressure plates in room
-function getUnpressedPressurePlates()
-  local currRoom = Game():GetRoom()
-  local pressurePlates = {}
-  local i = 1
-  while i < Game():GetRoom():GetGridSize() do
-    local gridEntity = currRoom:GetGridEntity(i)
-    if gridEntity ~= nil and gridEntity:GetType() == GridEntityType.GRID_PRESSURE_PLATE
-    and gridEntity:GetSaveState().State == 0 then
-      pressurePlates = append(pressurePlates, getGridIndex(gridEntity.Position)) 
-    end
-    i = i + 1
-  end
-  return pressurePlates
-end
-  
-
--- gets all entities in the room and puts them in a table list
-function getAllRoomEntities()
-  local entityList = Game():GetRoom():GetEntities()
-  
-  local entityGridList = {}
-  local listIndex = 1
-  local i = 0
-  local iterating = true
-  while iterating do
-    local entity = entityList:Get(i)
-    entityGridList[listIndex] = entity
-    listIndex = listIndex + 1
-    i = i + 1
-    if (i >= entityList:__len()) then
-      iterating = false
-    end
-  end
-  return entityGridList
-end
-
--- are there no enemies in this room?
-function noEnemies()
-  
-  local entityList = Game():GetRoom():GetEntities()
-  
-  local i = 0
-  while true do
-    local entity = entityList:Get(i)
-    if (entity:IsActiveEnemy()) then return false end
-    i = i + 1
-    if (i >= entityList:__len()) then
-      return true
-    end
-  end
-end
-
--- get a list of all game entities (not GRID entities) at the given grid index
-function getAllRoomEntitiesAtIndex(gridIndex)
-  local entityList = Game():GetRoom():GetEntities()
-  
-  local entityGridList = {}
-  local listIndex = 1
-  local i = 0
-  local iterating = true
-  while iterating do
-    local entity = entityList:Get(i)
-    if (getGridIndex(entity.Position)) == gridIndex then
-      entityGridList[listIndex] = entity
-      listIndex = listIndex + 1
-    end
-    i = i + 1
-    if (i >= entityList:__len()) then
-      iterating = false
-    end
-  end
-  return entityGridList
-end
-
--- returns a list of all grid indices of the pedestal items
-function getPedestalItems()
-  local entityGridList = {}
-  local allEntitiesList = getAllRoomEntities()
-  local listIndex = 1
-  for index, entity in pairs(allEntitiesList) do
-    if isPedestalItem(entity) then
-      entityGridList[listIndex] = entity
-      listIndex = listIndex + 1
-    end
-  end
-  return entityGridList
-end
-
--- returns a list of all grid indices of the pedestal items
-function getPassivePedestalItems()
-  return filter(isPassiveItem, getPedestalItems())
-end
-
--- returns the value that corresponds to what item it is
-function getPedestalItemId(pedestalItem)
-  return pedestalItem.SubType
-end
-
--- returns true if the given entity is a pedestal item
-function isPedestalItem(entity)
-  return entity.Type == 5 and entity.Variant == 100 and entity.SubType ~= 0
-end
-
-function isPassiveItem(pedestalItem)
-  local itemId = getPedestalItemId(pedestalItem)
-  
-  return (itemId == 320 or
-    itemId == 11 or
-    itemId == 245 or
-    itemId == 191 or
-    itemId == 116 or
-    itemId == 18 or
-    itemId == 132 or
-    itemId == 74 or
-    itemId == 346 or
-    itemId == 230 or
-    itemId == 188 or
-    itemId == 214 or
-    itemId == 161 or
-    itemId == 222 or
-    itemId == 308 or
-    itemId == 300 or
-    itemId == 207 or
-    itemId == 231 or
-    itemId == 272 or
-    itemId == 274 or
-    itemId == 247 or
-    itemId == 279 or
-    itemId == 260 or
-    itemId == 226 or
-    itemId == 119 or
-    itemId == 254 or
-    itemId == 7 or
-    itemId == 157 or
-    itemId == 342 or
-    itemId == 246 or
-    itemId == 273 or
-    itemId == 140 or
-    itemId == 125 or
-    itemId == 250 or
-    itemId == 131 or
-    itemId == 19 or
-    itemId == 198 or
-    itemId == 25 or
-    itemId == 118 or
-    itemId == 337 or
-    itemId == 8 or
-    itemId == 129 or
-    itemId == 144 or
-    itemId == 209 or
-    itemId == 340 or
-    itemId == 319 or
-    itemId == 301 or
-    itemId == 307 or
-    itemId == 165 or
-    itemId == 162 or
-    itemId == 216 or
-    itemId == 208 or
-    itemId == 62 or
-    itemId == 154 or
-    itemId == 69 or
-    itemId == 241 or
-    itemId == 224 or
-    itemId == 4 or
-    itemId == 73 or
-    itemId == 48 or
-    itemId == 316 or
-    itemId == 170 or
-    itemId == 278 or
-    itemId == 259 or
-    itemId == 117 or
-    itemId == 81 or
-    itemId == 185 or
-    itemId == 336 or
-    itemId == 237 or
-    itemId == 113 or
-    itemId == 24 or
-    itemId == 23 or
-    itemId == 57 or
-    itemId == 52 or
-    itemId == 265 or
-    itemId == 236 or
-    itemId == 168 or
-    itemId == 310 or
-    itemId == 240 or
-    itemId == 204 or
-    itemId == 179 or
-    itemId == 257 or
-    itemId == 128 or
-    itemId == 318 or
-    itemId == 163 or
-    itemId == 225 or
-    itemId == 210 or
-    itemId == 215 or
-    itemId == 331 or
-    itemId == 70 or
-    itemId == 112 or
-    itemId == 206 or
-    itemId == 212 or
-    itemId == 187 or
-    itemId == 134 or
-    itemId == 156 or
-    itemId == 10 or
-    itemId == 167 or
-    itemId == 269 or
-    itemId == 248 or
-    itemId == 184 or
-    itemId == 313 or
-    itemId == 178 or
-    itemId == 256 or
-    itemId == 203 or
-    itemId == 242 or
-    itemId == 148 or
-    itemId == 234 or
-    itemId == 149 or
-    itemId == 201 or
-    itemId == 276 or
-    itemId == 197 or
-    itemId == 311 or
-    itemId == 266 or
-    itemId == 238 or
-    itemId == 239 or
-    itemId == 343 or
-    itemId == 332 or
-    itemId == 270 or
-    itemId == 302 or
-    itemId == 15 or
-    itemId == 304 or
-    itemId == 275 or
-    itemId == 277 or
-    itemId == 252 or
-    itemId == 96 or
-    itemId == 88 or
-    itemId == 99 or
-    itemId == 100 or
-    itemId == 87 or
-    itemId == 82 or
-    itemId == 213 or
-    itemId == 46 or
-    itemId == 22 or
-    itemId == 312 or
-    itemId == 194 or
-    itemId == 12 or
-    itemId == 253 or
-    itemId == 53 or
-    itemId == 344 or
-    itemId == 193 or
-    itemId == 202 or
-    itemId == 71 or
-    itemId == 258 or
-    itemId == 262 or
-    itemId == 173 or
-    itemId == 195 or
-    itemId == 110 or
-    itemId == 55 or
-    itemId == 200 or
-    itemId == 30 or
-    itemId == 199 or
-    itemId == 114 or
-    itemId == 31 or
-    itemId == 228 or
-    itemId == 139 or
-    itemId == 29 or
-    itemId == 217 or
-    itemId == 109 or
-    itemId == 322 or
-    itemId == 229 or
-    itemId == 106 or
-    itemId == 153 or
-    itemId == 5 or
-    itemId == 317 or
-    itemId == 271 or
-    itemId == 6 or
-    itemId == 121 or
-    itemId == 120 or
-    itemId == 219 or
-    itemId == 115 or
-    itemId == 141 or
-    itemId == 51 or
-    itemId == 75 or
-    itemId == 227 or
-    itemId == 309 or
-    itemId == 218 or
-    itemId == 169 or
-    itemId == 261 or
-    itemId == 281 or
-    itemId == 190 or
-    itemId == 223 or
-    itemId == 174 or
-    itemId == 16 or
-    itemId == 95 or
-    itemId == 267 or
-    itemId == 14 or
-    itemId == 72 or
-    itemId == 268 or
-    itemId == 26 or
-    itemId == 221 or
-    itemId == 94 or
-    itemId == 182 or
-    itemId == 172 or
-    itemId == 220 or
-    itemId == 339 or
-    itemId == 306 or
-    itemId == 321 or
-    itemId == 142 or
-    itemId == 305 or
-    itemId == 255 or
-    itemId == 205 or
-    itemId == 280 or
-    itemId == 67 or
-    itemId == 9 or
-    itemId == 17 or
-    itemId == 264 or
-    itemId == 189 or
-    itemId == 330 or
-    itemId == 143 or
-    itemId == 91 or
-    itemId == 211 or
-    itemId == 89 or
-    itemId == 159 or
-    itemId == 3 or
-    itemId == 196 or
-    itemId == 251 or
-    itemId == 64 or
-    itemId == 176 or
-    itemId == 50 or
-    itemId == 138 or
-    itemId == 232 or
-    itemId == 315 or
-    itemId == 92 or
-    itemId == 345 or
-    itemId == 299 or
-    itemId == 244 or
-    itemId == 68 or
-    itemId == 152 or
-    itemId == 63 or
-    itemId == 28 or
-    itemId == 180 or
-    itemId == 334 or
-    itemId == 103 or
-    itemId == 21 or
-    itemId == 101 or
-    itemId == 2 or
-    itemId == 60 or
-    itemId == 329 or
-    itemId == 79 or
-    itemId == 333 or
-    itemId == 151 or
-    itemId == 328 or
-    itemId == 80 or
-    itemId == 104 or
-    itemId == 155 or
-    itemId == 327 or
-    itemId == 98 or
-    itemId == 1 or
-    itemId == 90 or
-    itemId == 335 or
-    itemId == 13 or
-    itemId == 108 or
-    itemId == 249 or
-    itemId == 314 or
-    itemId == 233 or
-    itemId == 183 or
-    itemId == 341 or
-    itemId == 150 or
-    itemId == 20 or
-    itemId == 54 or
-    itemId == 243 or
-    itemId == 303 or
-    itemId == 122 or
-    itemId == 32 or
-    itemId == 27 or
-    itemId == 76 or
-    itemId == 359 or
-    itemId == 408 or
-    itemId == 391 or
-    itemId == 438 or
-    itemId == 420 or
-    itemId == 353 or
-    itemId == 385 or
-    itemId == 377 or
-    itemId == 412 or
-    itemId == 356 or
-    itemId == 387 or
-    itemId == 402 or
-    itemId == 372 or
-    itemId == 423 or
-    itemId == 369 or
-    itemId == 354 or
-    itemId == 415 or
-    itemId == 371 or
-    itemId == 373 or
-    itemId == 416 or
-    itemId == 381 or
-    itemId == 409 or
-    itemId == 368 or
-    itemId == 410 or
-    itemId == 401 or
-    itemId == 404 or
-    itemId == 361 or
-    itemId == 364 or
-    itemId == 418 or
-    itemId == 405 or
-    itemId == 432 or
-    itemId == 398 or
-    itemId == 429 or
-    itemId == 374 or
-    itemId == 375 or
-    itemId == 413 or
-    itemId == 360 or
-    itemId == 388 or
-    itemId == 440 or
-    itemId == 362 or
-    itemId == 384 or
-    itemId == 435 or
-    itemId == 365 or
-    itemId == 411 or
-    itemId == 394 or
-    itemId == 399 or
-    itemId == 436 or
-    itemId == 355 or
-    itemId == 414 or
-    itemId == 370 or
-    itemId == 431 or
-    itemId == 433 or
-    itemId == 425 or
-    itemId == 378 or
-    itemId == 426 or
-    itemId == 430 or
-    itemId == 380 or
-    itemId == 428 or
-    itemId == 379 or
-    itemId == 407 or
-    itemId == 376 or
-    itemId == 389 or
-    itemId == 424 or
-    itemId == 366 or
-    itemId == 390 or
-    itemId == 393 or
-    itemId == 400 or
-    itemId == 403 or
-    itemId == 367 or
-    itemId == 417 or
-    itemId == 363 or
-    itemId == 395 or
-    itemId == 358 or
-    itemId == 350 or
-    itemId == 397 or
-    itemId == 392 or
-    itemId == 526 or
-    itemId == 491 or
-    itemId == 493 or
-    itemId == 465 or
-    itemId == 528 or
-    itemId == 511 or
-    itemId == 443 or
-    itemId == 506 or
-    itemId == 458 or
-    itemId == 473 or
-    itemId == 535 or
-    itemId == 509 or
-    itemId == 513 or
-    itemId == 549 or
-    itemId == 514 or
-    itemId == 551 or
-    itemId == 518 or
-    itemId == 497 or
-    itemId == 453 or
-    itemId == 457 or
-    itemId == 466 or
-    itemId == 455 or
-    itemId == 546 or
-    itemId == 442 or
-    itemId == 446 or
-    itemId == 530 or
-    itemId == 469 or
-    itemId == 547 or
-    itemId == 445 or
-    itemId == 498 or
-    itemId == 499 or
-    itemId == 496 or
-    itemId == 462 or
-    itemId == 450 or
-    itemId == 517 or
-    itemId == 467 or
-    itemId == 540 or
-    itemId == 495 or
-    itemId == 460 or
-    itemId == 464 or
-    itemId == 501 or
-    itemId == 531 or
-    itemId == 543 or
-    itemId == 470 or
-    itemId == 494 or
-    itemId == 548 or
-    itemId == 520 or
-    itemId == 472 or
-    itemId == 532 or
-    itemId == 502 or
-    itemId == 444 or
-    itemId == 525 or
-    itemId == 519 or
-    itemId == 471 or
-    itemId == 537 or
-    itemId == 447 or
-    itemId == 503 or
-    itemId == 538 or
-    itemId == 541 or
-    itemId == 449 or
-    itemId == 456 or
-    itemId == 508 or
-    itemId == 539 or
-    itemId == 461 or
-    itemId == 544 or
-    itemId == 505 or
-    itemId == 454 or
-    itemId == 529 or
-    itemId == 500 or
-    itemId == 534 or
-    itemId == 468 or
-    itemId == 448 or
-    itemId == 459 or
-    itemId == 542 or
-    itemId == 463 or
-    itemId == 451 or
-    itemId == 524 or
-    itemId == 474 or
-    itemId == 533 or
-    itemId == 452 or
-    itemId == 492)
-  
-end
-
--- returns whether an entity blocks your movement or not
--- should avoid these ones in the roomSearch problem
-function entityBlocksMovement(entity)
-  return (entity.Type == 5 and entity.Variant == 100) -- pedestal items
-  or (entity.Type == 6) -- slot machines and bums
-end
-
--- returns a string representing the enum type of the given game entity
-function getEntityType(entity)
-  
-  local entityType = entity.Type
-  local entityVariant = entity.Variant
-  
-  -- weird misc type that the game uses for random entities
-  if entityType == 1000 then return "ENTITY_MISC" end
-  
-  -- filter bogus input
-  if entityType < 0 or entityType > 225 then return nil end
-  
-  -- match type
-  if entityType == 0   then return "ENTITY_NULL" end
-  if entityType == 1   then return "ENTITY_PLAYER" end
-  if entityType == 2   then return "ENTITY_TEAR" end
-  if entityType == 3   then return "ENTITY_FAMILIAR" end
-  if entityType == 4   then return "ENTITY_BOMBDROP" end
-  
-  -- there's a lot of different pickups lol
-  if entityType == 5 then
-    local pickupString = "ENTITY_PICKUP"
-    local variantString = ""
-    
-    -- define variants
-    if entityVariant == 20 then
-      variantString = "COIN"
-    elseif entityVariant == 40 then
-      variantString = "BOMB"
-    elseif entityVariant == 60 then
-      variantString = "CHEST"
-    elseif entityVariant == 69 then
-      variantString = "BAG"
-    elseif entityVariant == 30 then
-      variantString = "KEY"
-    elseif entityVariant == 100 then
-      variantString = "PEDESTAL_ITEM"
-    elseif entityVariant == 10 then
-      variantString = "HEART"
-    elseif entityVariant == 10 then
-      variantString = "BUM"
-    elseif entityVariant == 10 then
-      variantString = "LOTTERY"
-    end
-    
-    -- return string
-    if variantString == "" then
-      return pickupString
-    else
-      return pickupString .. "_" .. variantString
-    end
-  end
-  
-  if entityType == 6   then return "ENTITY_SLOT" end
-  if entityType == 7   then return "ENTITY_LASER" end
-  if entityType == 8   then return "ENTITY_KNIFE" end
-  if entityType == 9   then return "ENTITY_PROJECTILE" end
-  if entityType == 10  then return "ENTITY_GAPER" end
-  if entityType == 11  then return "ENTITY_GUSHER" end
-  if entityType == 12  then return "ENTITY_HORF" end
-  if entityType == 13  then return "ENTITY_FLY" end
-  if entityType == 14  then return "ENTITY_POOTER" end
-  if entityType == 15  then return "ENTITY_CLOTTY" end
-  if entityType == 16  then return "ENTITY_MULLIGAN" end
-  if entityType == 17  then return "ENTITY_SHOPKEEPER" end
-  if entityType == 18  then return "ENTITY_ATTACKFLY" end
-  if entityType == 19  then return "ENTITY_LARRYJR" end
-  if entityType == 20  then return "ENTITY_MONSTRO" end
-  if entityType == 21  then return "ENTITY_MAGGOT" end
-  if entityType == 22  then return "ENTITY_HIVE" end
-  if entityType == 23  then return "ENTITY_CHARGER" end
-  if entityType == 24  then return "ENTITY_GLOBIN" end
-  if entityType == 25  then return "ENTITY_BOOMFLY" end
-  if entityType == 26  then return "ENTITY_MAW" end
-  if entityType == 27  then return "ENTITY_HOST" end
-  if entityType == 28  then return "ENTITY_CHUB" end
-  if entityType == 29  then return "ENTITY_HOPPER" end
-  if entityType == 30  then return "ENTITY_BOIL" end
-  if entityType == 31  then return "ENTITY_SPITY" end
-  if entityType == 32  then return "ENTITY_BRAIN" end
-  if entityType == 33  then return "ENTITY_FIREPLACE" end
-  if entityType == 34  then return "ENTITY_LEAPER" end
-  if entityType == 35  then return "ENTITY_MRMAW" end
-  if entityType == 36  then return "ENTITY_GURDY" end
-  if entityType == 37  then return "ENTITY_BABY" end
-  if entityType == 38  then return "ENTITY_VIS" end
-  if entityType == 39  then return "ENTITY_GUTS" end
-  if entityType == 40  then return "ENTITY_KNIGHT" end
-  if entityType == 41  then return "ENTITY_STONEHEAD" end
-  if entityType == 42  then return "ENTITY_MONSTRO2" end
-  if entityType == 43  then return "ENTITY_POKY" end
-  if entityType == 44  then return "ENTITY_MOM" end
-  if entityType == 45  then return "ENTITY_SLOTH" end
-  if entityType == 46  then return "ENTITY_LUST" end
-  if entityType == 47  then return "ENTITY_WRATH" end
-  if entityType == 48  then return "ENTITY_GLUTTONY" end
-  if entityType == 49  then return "ENTITY_GREED" end
-  if entityType == 50  then return "ENTITY_ENVY" end
-  if entityType == 51  then return "ENTITY_PRIDE" end
-  if entityType == 52  then return "ENTITY_DOPLE" end
-  if entityType == 53  then return "ENTITY_FLAMINGHOPPER" end
-  if entityType == 54  then return "ENTITY_LEECH" end
-  if entityType == 55  then return "ENTITY_LUMP" end
-  if entityType == 56  then return "ENTITY_MEMBRAIN" end
-  if entityType == 57  then return "ENTITY_PARA_BITE" end
-  if entityType == 58  then return "ENTITY_FRED" end
-  if entityType == 59  then return "ENTITY_EYE" end
-  if entityType == 60  then return "ENTITY_SUCKER" end
-  if entityType == 61  then return "ENTITY_PIN" end
-  if entityType == 62  then return "ENTITY_FAMINE" end
-  if entityType == 63  then return "ENTITY_PESTILENCE" end
-  if entityType == 64  then return "ENTITY_WAR" end
-  if entityType == 65  then return "ENTITY_DEATH" end
-  if entityType == 66  then return "ENTITY_DUKE" end
-  if entityType == 67  then return "ENTITY_PEEP" end
-  if entityType == 68  then return "ENTITY_LOKI" end
-  if entityType == 69  then return "ENTITY_FISTULA_BIG" end
-  if entityType == 70  then return "ENTITY_FISTULA_MEDIUM" end
-  if entityType == 71  then return "ENTITY_FISTULA_SMALL" end
-  if entityType == 72  then return "ENTITY_BLASTOCYST_BIG" end
-  if entityType == 73  then return "ENTITY_BLASTOCYST_MEDIUM" end
-  if entityType == 74  then return "ENTITY_BLASTOCYST_SMALL" end
-  if entityType == 75  then return "ENTITY_EMBRYO" end
-  if entityType == 76  then return "ENTITY_MOMS_HEART" end
-  if entityType == 77  then return "ENTITY_GEMINI" end
-  if entityType == 78  then return "ENTITY_MOTER" end
-  if entityType == 79  then return "ENTITY_FALLEN" end
-  if entityType == 80  then return "ENTITY_HEADLESS_HORSEMAN" end
-  if entityType == 81  then return "ENTITY_HORSEMAN_HEAD" end
-  if entityType == 82  then return "ENTITY_SATAN" end
-  if entityType == 83  then return "ENTITY_SPIDER" end
-  if entityType == 84  then return "ENTITY_KEEPER" end
-  if entityType == 85  then return "ENTITY_GURGLE" end
-  if entityType == 86  then return "ENTITY_WALKINGBOIL" end
-  if entityType == 87  then return "ENTITY_BUTTLICKER" end
-  if entityType == 88  then return "ENTITY_HANGER" end
-  if entityType == 89  then return "ENTITY_SWARMER" end
-  if entityType == 90  then return "ENTITY_HEART" end
-  if entityType == 91  then return "ENTITY_MASK" end
-  if entityType == 92  then return "ENTITY_BIGSPIDER" end
-  if entityType == 93  then return "ENTITY_ETERNALFLY" end
-  if entityType == 94  then return "ENTITY_MASK_OF_INFAMY" end
-  if entityType == 95  then return "ENTITY_HEART_OF_INFAMY" end
-  if entityType == 96  then return "ENTITY_GURDY_JR" end
-  if entityType == 97  then return "ENTITY_WIDOW" end
-  if entityType == 98  then return "ENTITY_DADDYLONGLEGS" end
-  if entityType == 99  then return "ENTITY_ISAAC" end
-  if entityType == 100 then return "ENTITY_STONE_EYE" end
-  if entityType == 101 then return "ENTITY_CONSTANT_STONE_SHOOTER" end
-  if entityType == 102 then return "ENTITY_BRIMSTONE_HEAD" end
-  if entityType == 103 then return "ENTITY_MOBILE_HOST" end
-  if entityType == 104 then return "ENTITY_NEST" end
-  if entityType == 105 then return "ENTITY_BABY_LONG_LEGS" end
-  if entityType == 106 then return "ENTITY_CRAZY_LONG_LEGS" end
-  if entityType == 107 then return "ENTITY_FATTY" end
-  if entityType == 108 then return "ENTITY_FAT_SACK" end
-  if entityType == 109 then return "ENTITY_BLUBBER" end
-  if entityType == 110 then return "ENTITY_HALF_SACK" end
-  if entityType == 111 then return "ENTITY_DEATHS_HEAD" end
-  if entityType == 112 then return "ENTITY_MOMS_HAND" end
-  if entityType == 113 then return "ENTITY_FLY_L2" end
-  if entityType == 114 then return "ENTITY_SPIDER_L2" end
-  if entityType == 115 then return "ENTITY_SWINGER" end
-  if entityType == 116 then return "ENTITY_DIP" end
-  if entityType == 117 then return "ENTITY_WALL_HUGGER" end
-  if entityType == 118 then return "ENTITY_WIZOOB" end
-  if entityType == 119 then return "ENTITY_SQUIRT" end
-  if entityType == 120 then return "ENTITY_COD_WORM" end
-  if entityType == 121 then return "ENTITY_RING_OF_FLIES" end
-  if entityType == 122 then return "ENTITY_DINGA" end
-  if entityType == 123 then return "ENTITY_OOB" end
-  if entityType == 124 then return "ENTITY_BLACK_MAW" end
-  if entityType == 125 then return "ENTITY_SKINNY" end
-  if entityType == 126 then return "ENTITY_BONY" end
-  if entityType == 127 then return "ENTITY_HOMUNCULUS" end
-  if entityType == 128 then return "ENTITY_TUMOR" end
-  if entityType == 129 then return "ENTITY_CAMILLO_JR" end
-  if entityType == 130 then return "ENTITY_NERVE_ENDING" end
-  if entityType == 131 then return "ENTITY_SKINBALL" end
-  if entityType == 132 then return "ENTITY_MOM_HEAD" end
-  if entityType == 133 then return "ENTITY_ONE_TOOTH" end
-  if entityType == 134 then return "ENTITY_GAPING_MAW" end
-  if entityType == 135 then return "ENTITY_BROKEN_GAPING_MAW" end
-  if entityType == 136 then return "ENTITY_GURGLING" end
-  if entityType == 137 then return "ENTITY_SPLASHER" end
-  if entityType == 138 then return "ENTITY_GRUB" end
-  if entityType == 139 then return "ENTITY_WALL_CREEP" end
-  if entityType == 140 then return "ENTITY_RAGE_CREEP" end
-  if entityType == 141 then return "ENTITY_BLIND_CREEP" end
-  if entityType == 142 then return "ENTITY_CONJOINED_SPITTY" end
-  if entityType == 143 then return "ENTITY_ROUND_WORM" end
-  if entityType == 144 then return "ENTITY_POOP" end
-  if entityType == 145 then return "ENTITY_RAGLING" end
-  if entityType == 146 then return "ENTITY_FLESH_MOBILE_HOST" end
-  if entityType == 147 then return "ENTITY_PSY_HORF" end
-  if entityType == 148 then return "ENTITY_FULL_FLY" end
-  if entityType == 149 then return "ENTITY_TICKING_SPIDER" end
-  if entityType == 150 then return "ENTITY_BEGOTTEN" end
-  if entityType == 151 then return "ENTITY_NULLS" end
-  if entityType == 152 then return "ENTITY_PSY_TUMOR" end
-  if entityType == 153 then return "ENTITY_FLOATING_KNIGHT" end
-  if entityType == 154 then return "ENTITY_NIGHT_CRAWLER" end
-  if entityType == 155 then return "ENTITY_DART_FLY" end
-  if entityType == 156 then return "ENTITY_CONJOINED_FATTY" end
-  if entityType == 157 then return "ENTITY_FAT_BAT" end
-  if entityType == 158 then return "ENTITY_IMP" end
-  if entityType == 159 then return "ENTITY_THE_HAUNT" end
-  if entityType == 160 then return "ENTITY_DINGLE" end
-  if entityType == 161 then return "ENTITY_MEGA_MAW" end
-  if entityType == 162 then return "ENTITY_GATE" end
-  if entityType == 163 then return "ENTITY_MEGA_FATTY" end
-  if entityType == 164 then return "ENTITY_CAGE" end
-  if entityType == 165 then return "ENTITY_MAMA_GURDY" end
-  if entityType == 166 then return "ENTITY_DARK_ONE" end
-  if entityType == 167 then return "ENTITY_ADVERSARY" end
-  if entityType == 168 then return "ENTITY_POLYCEPHALUS" end
-  if entityType == 169 then return "ENTITY_MR_FRED" end
-  if entityType == 170 then return "ENTITY_URIEL" end
-  if entityType == 171 then return "ENTITY_GABRIEL" end
-  if entityType == 172 then return "ENTITY_THE_LAMB" end
-  if entityType == 173 then return "ENTITY_MEGA_SATAN" end
-  if entityType == 174 then return "ENTITY_MEGA_SATAN_2" end
-  if entityType == 175 then return "ENTITY_ROUNDY" end
-  if entityType == 176 then return "ENTITY_BLACK_BONY" end
-  if entityType == 177 then return "ENTITY_BLACK_GLOBIN" end
-  if entityType == 178 then return "ENTITY_BLACK_GLOBIN_HEAD" end
-  if entityType == 179 then return "ENTITY_BLACK_GLOBIN_BODY" end
-  if entityType == 180 then return "ENTITY_SWARM" end
-  if entityType == 181 then return "ENTITY_MEGA_CLOTTY" end
-  if entityType == 182 then return "ENTITY_BONE_KNIGHT" end
-  if entityType == 183 then return "ENTITY_CYCLOPIA" end
-  if entityType == 184 then return "ENTITY_RED_GHOST" end
-  if entityType == 185 then return "ENTITY_FLESH_DEATHS_HEAD" end
-  if entityType == 186 then return "ENTITY_MOMS_DEAD_HAND" end
-  if entityType == 187 then return "ENTITY_DUKIE" end
-  if entityType == 188 then return "ENTITY_ULCER" end
-  if entityType == 189 then return "ENTITY_MEATBALL" end
-  if entityType == 190 then return "ENTITY_PITFALL" end
-  if entityType == 191 then return "ENTITY_MOVABLE_TNT" end
-  if entityType == 192 then return "ENTITY_ULTRA_COIN" end
-  if entityType == 193 then return "ENTITY_ULTRA_DOOR" end
-  if entityType == 194 then return "ENTITY_CORN_MINE" end
-  if entityType == 195 then return "ENTITY_HUSH_FLY" end
-  if entityType == 196 then return "ENTITY_HUSH_GAPER" end
-  if entityType == 197 then return "ENTITY_HUSH_BOIL" end
-  if entityType == 198 then return "ENTITY_GREED_GAPER" end
-  if entityType == 199 then return "ENTITY_MUSHROOM" end
-  if entityType == 200 then return "ENTITY_POISON_MIND" end
-  if entityType == 201 then return "ENTITY_STONEY" end
-  if entityType == 202 then return "ENTITY_BLISTER" end
-  if entityType == 203 then return "ENTITY_THE_THING" end
-  if entityType == 204 then return "ENTITY_MINISTRO" end
-  if entityType == 205 then return "ENTITY_PORTAL" end
-  if entityType == 206 then return "ENTITY_TARBOY" end
-  if entityType == 207 then return "ENTITY_FISTULOID" end
-  if entityType == 208 then return "ENTITY_GUSH" end
-  if entityType == 209 then return "ENTITY_LEPER" end
-  if entityType == 210 then return "ENTITY_STAIN" end
-  if entityType == 211 then return "ENTITY_BROWNIE" end
-  if entityType == 212 then return "ENTITY_FORSAKEN" end
-  if entityType == 213 then return "ENTITY_LITTLE_HORN" end
-  if entityType == 214 then return "ENTITY_RAG_MAN" end
-  if entityType == 215 then return "ENTITY_ULTRA_GREED" end
-  if entityType == 216 then return "ENTITY_HUSH" end
-  if entityType == 217 then return "ENTITY_HUSH_SKINLESS" end
-  if entityType == 218 then return "ENTITY_RAG_MEGA" end
-  if entityType == 219 then return "ENTITY_SISTERS_VIS" end
-  if entityType == 220 then return "ENTITY_BIG_HORN" end
-  if entityType == 221 then return "ENTITY_DELIRIUM" end
-  if entityType == 222 then return "ENTITY_MATRIARCH" end
-  if entityType == 223 then return "ENTITY_EFFECT" end
-  if entityType == 224 then return "ENTITY_TEXT" end
-  return nil -- returns nil if it doesn't match a type
-end
-
-function printAdjacentGridIndices()
-  local yOffset = 10
-  local n = getAllAdjacentGridIndices(getPlayerGridIndex())
-  local screenPos = getPlayerScreenPosition()
-  currentIndex = tostring(getPlayerGridIndex())
-  upIndex = tostring(n[1])
-  downIndex = tostring(n[2])
-  leftIndex = tostring(n[3])
-  rightIndex = tostring(n[4])
-  upIsBlocked = isGridIndexBlocked(n[1])
-  downIsBlocked = isGridIndexBlocked(n[2])
-  leftIsBlocked = isGridIndexBlocked(n[3])
-  rightIsBlocked = isGridIndexBlocked(n[4])
-  Isaac.RenderText(currentIndex, screenPos.X - string.len(currentIndex) * 3, screenPos.Y - yOffset, 1, 1, 1, 1)
-  if (upIsBlocked) then
-    Isaac.RenderText(upIndex .. ": " .. tostring(getGridType(n[1])), screenPos.X - string.len(upIndex) * 3, screenPos.Y - yOffset - 30, 1, 0, 0, 1)
-  else
-    Isaac.RenderText(upIndex, screenPos.X - string.len(upIndex) * 3, screenPos.Y - yOffset - 30, 1, 1, 1, 1)
-  end
-  if (downIsBlocked) then
-    Isaac.RenderText(downIndex .. ": " .. tostring(getGridType(n[2])), screenPos.X - string.len(downIndex) * 3, screenPos.Y - yOffset + 30, 1, 0, 0, 1)
-  else
-    Isaac.RenderText(downIndex, screenPos.X - string.len(downIndex) * 3, screenPos.Y - yOffset + 30, 1, 1, 1, 1)
-  end
-  if (leftIsBlocked) then
-    Isaac.RenderText(tostring(getGridType(n[3])) .. ": " .. leftIndex, screenPos.X - string.len(getGridType(n[2]) .. ": " .. leftIndex) * 5 - 30, screenPos.Y - yOffset, 1, 0, 0, 1)
-  else
-    Isaac.RenderText(leftIndex, screenPos.X - string.len(leftIndex) * 3 - 30, screenPos.Y - yOffset, 1, 1, 1, 1)
-  end
-  if (rightIsBlocked) then
-    Isaac.RenderText(rightIndex .. ": " .. tostring(getGridType(n[4])), screenPos.X - string.len(rightIndex) * 3 + 30, screenPos.Y - yOffset, 1, 0, 0, 1)
-  else
-    Isaac.RenderText(rightIndex, screenPos.X - string.len(rightIndex) * 3 + 30, screenPos.Y - yOffset, 1, 1, 1, 1)
-  end
-end
-
-function aStarRoomSearch(index1, index2)
-  local cost = 1
-  local goalIndex = index2 -- our goal is to get from pos1 to pos2
-  local openNodes = {}
-  local closedNodes = {}
-  local g = {}
-  local pq  = PriorityQueue()
-  
-  local initialNode = {index1, {}} -- a tuple of value and path
-  pq:put(initialNode, 0)
-  g[index1] = 0
-  
-  while (not pq:empty()) do
-    local topOfQueue = pq:pop()
-    local currentIndex = topOfQueue[1]
-    local pathToIndex = topOfQueue[2]
-    -- goal test
-    if (currentIndex == goalIndex) then
-      return pathToIndex
-    end
-    if (not closedNodes[currentIndex] == true) then
-      closedNodes[currentIndex] = true
-      for indexInList, nextIndex in pairs(nextValidGridIndices(currentIndex, goalIndex)) do
-        costToNextIndex = g[currentIndex] + cost + manhattanDist(nextIndex, goalIndex)
-        if (not closedNodes[nextIndex] == true) then
-          if (not openNodes[nextIndex] == true) then
-             openNodes[nextIndex] = true
-             pq:put({nextIndex, append(pathToIndex, nextIndex)}, costToNextIndex)
-             g[nextIndex] = g[currentIndex] + cost
-          end
-        else
-          if (costToNextIndex < g[nextIndex]) then
-            -- pq:put({nextIndex, append(pathToIndex, nextIndex)}, costToNextIndex) -- update, not put
-            g[nextIndex] = g[currentIndex] + cost
-            openNodes[nextIndex] = false
-          end
-        end
-      end
-    end
-  end
-  -- if there is no path between them, then return an empty set of directions
-  return {}
-end
-
-function aStarToPos(pos)
-  return aStarRoomSearch(getGridIndex(getPlayerPosition()), getGridIndex(pos))
-end
-
-function convertListOfIndexToPos(indexList)
-  local convertedList = {}
-  for indexInList, nextIndex in pairs(indexList) do
-    convertedList[indexInList] = getGridPos(nextIndex)
-  end
-  return convertedList
-end
-
--- returns a list of Vector indicating the directions to get to the given pos
-function getDirectionsTo(pos)
-  return convertListOfIndexToPos(aStarToPos(pos))
-end
-
-function printAllGridIndices(listOfDirections)
-  for indexInList, nextPos in pairs(listOfDirections) do
-    local screenPos = getScreenPosition(nextPos)
-    Isaac.RenderText(tostring(getGridIndex(nextPos)), screenPos.X, screenPos.Y, 1, 1, 1, 1)
-  end
-end
-
-function printAllGameEntities(listOfEntities)
-  for indexInList, entity in pairs(listOfEntities) do
-    local screenPos = getScreenPosition(entity.Position)
-    local entityString = tostring(getEntityType(entity))
-    Isaac.RenderText(entityString, screenPos.X - string.len(entityString) * 3, screenPos.Y, 1, 1, 1, 1)
-  end
-end
-
-function printNaiveDFS()
-  local visitedRoomString = "Visited Rooms: "
-  for room, hasVisited in pairs(visitedRooms2) do
-    visitedRoomString = visitedRoomString .. tostring(room) .. ", "
-  end
-  Isaac.RenderText(visitedRoomString, 10, 80, 1, 1, 1, 1)
-end
-
-function printDFS()
-  local currentRoomIndex = "currentRoom = " .. tostring(getCurrentRoom())
-  Isaac.RenderText("LEVEL SEARCH", 100, 70, 1, 1, 1, 1)
-  Isaac.RenderText(currentRoomIndex, 100, 100, 1, 1, 1, 1)
-  local topOfRoomStack = "top of RoomStack = " .. tostring(roomStack:peek()[1])
-  Isaac.RenderText(topOfRoomStack, 100, 130, 1, 1, 1, 1)
-end
-
--- code to be run every frame
-function onStep()
-  -- enable and disable the AI mod by pressing 'R' on your keyboard
-  if Input.IsActionTriggered(ButtonAction.ACTION_RESTART, 0) then
-    if modEnabled then
-      modEnabled = false
-      Isaac.ConsoleOutput("AI Mod Disabled\n")
-      setIsaacMessage("AI Disabled (" .. agentTypeString .. ")", 100)
-    else
-      modEnabled = true
-      timer = 0
-      Isaac.ConsoleOutput("AI Mod Enabled\n")
-      setIsaacMessage("AI Enabled (" .. agentTypeString .. ")", 100)
-    end
-  end
-  
-  -- print the isaacMessage
-  printIsaacMessage()
-  
-  if modEnabled then
-    -- this agent moves left and then right every moveLeftAndRightEvery tics
-    if agentType == AgentType.MoveLeftAndRight then
-      if timer % moveLeftAndRightEvery == 0 then
-        if moveDirectionX == ButtonAction.ACTION_LEFT then
-          moveDirectionX = ButtonAction.ACTION_RIGHT
-        else
-          moveDirectionX = ButtonAction.ACTION_LEFT
-        end
-        timer = 0
-      end
-      timer = timer + 1
-    end
-    
-    -- this agent moves and shoots in the last move / shoot direction
-    if agentType == AgentType.SnakeAgent then
-      if Input.IsActionTriggered(ButtonAction.ACTION_LEFT, 0) or Input.IsActionTriggered(ButtonAction.ACTION_SHOOTLEFT, 0) then
-        moveDirectionX = ButtonAction.ACTION_LEFT
-        moveDirectionY = nil
-        shootDirection = ButtonAction.ACTION_SHOOTRIGHT
-      elseif Input.IsActionTriggered(ButtonAction.ACTION_RIGHT, 0) or Input.IsActionTriggered(ButtonAction.ACTION_SHOOTRIGHT, 0) then
-        moveDirectionX = ButtonAction.ACTION_RIGHT
-        moveDirectionY = nil
-        shootDirection = ButtonAction.ACTION_SHOOTLEFT
-      elseif Input.IsActionTriggered(ButtonAction.ACTION_UP, 0) or Input.IsActionTriggered(ButtonAction.ACTION_SHOOTUP, 0) then
-        moveDirectionX = nil
-        moveDirectionY = ButtonAction.ACTION_UP
-        shootDirection = ButtonAction.ACTION_SHOOTDOWN
-      elseif Input.IsActionTriggered(ButtonAction.ACTION_DOWN, 0) or Input.IsActionTriggered(ButtonAction.ACTION_SHOOTDOWN, 0) then
-        moveDirectionX = nil
-        moveDirectionY = ButtonAction.ACTION_DOWN
-        shootDirection = ButtonAction.ACTION_SHOOTUP
-      end
-    end
-    
-    -- this agent moves directly to the point on the screen that you click!
-    if agentType == AgentType.DumbPointAndClick then
-      shootDirection = nil
-      moveDirectionX = nil
-      moveDirectionY = nil
-      if Input.IsMouseBtnPressed(0) then
-        pointAndClickPos = Input.GetMousePosition(true)
-      end
-      if pointAndClickPos ~= nil then
-        local mousePosScreen = Isaac.WorldToScreen(pointAndClickPos)
-        Isaac.RenderText("X", mousePosScreen.X - 3, mousePosScreen.Y - 6, 1, 0, 0, 1)
-        
-        local playerPos =  getPlayerPosition()
-        
-        local xDistToClickPos = pointAndClickPos.X - playerPos.X
-        local yDistToClickPos = pointAndClickPos.Y - playerPos.Y
-        
-        if math.abs(xDistToClickPos) > pointAndClickThreshold then
-          if xDistToClickPos > 0 then
-            moveDirectionX = ButtonAction.ACTION_RIGHT
-            
-          elseif xDistToClickPos < 0 then
-            moveDirectionX = ButtonAction.ACTION_LEFT
-          end
-        end
-        
-        if math.abs(yDistToClickPos) > pointAndClickThreshold then
-          if yDistToClickPos < 0 then
-            moveDirectionY = ButtonAction.ACTION_UP
-            
-          elseif yDistToClickPos > 0 then
-            moveDirectionY = ButtonAction.ACTION_DOWN
-          end
-        end
-      end
-    end
-    
-    -- this agent moves directly to the point on the screen that you click!
-    if agentType == AgentType.SmartPointAndClick then
-      shootDirection = nil
-      moveDirectionX = nil
-      moveDirectionY = nil
-      
-      if Input.IsMouseBtnPressed(0) then
-        pointAndClickPos = Input.GetMousePosition(true)
-        directions = getDirectionsTo(pointAndClickPos)
-        directionIndex = 1
-      end
-      if pointAndClickPos ~= nil and directions and directionIndex <= #directions then
-        local mousePosScreen = Isaac.WorldToScreen(pointAndClickPos)
-        Isaac.RenderText("X", mousePosScreen.X - 3, mousePosScreen.Y - 6, 1, 0, 0, 1)
-        
-        -- print all of the grid indexes at their positions
-        printAllGridIndices(directions)
-        
-        local playerPos =  getPlayerPosition()
-        
-        local xDistToNextPos = directions[directionIndex].X - playerPos.X
-        local yDistToNextPos = directions[directionIndex].Y - playerPos.Y
-        
-        if math.abs(xDistToNextPos) > pointAndClickThreshold then
-          if xDistToNextPos > 0 then
-            moveDirectionX = ButtonAction.ACTION_RIGHT
-            
-          elseif xDistToNextPos < 0 then
-            moveDirectionX = ButtonAction.ACTION_LEFT
-          end
-        end
-        
-        if math.abs(yDistToNextPos) > pointAndClickThreshold then
-          if yDistToNextPos < 0 then
-            moveDirectionY = ButtonAction.ACTION_UP
-            
-          elseif yDistToNextPos > 0 then
-            moveDirectionY = ButtonAction.ACTION_DOWN
-          end
-        end
-        
-        -- move on to next position if we reach this one, unless we are forcing to move to the goal
-        if math.abs(xDistToNextPos) < pointAndClickThreshold and
-        math.abs(yDistToNextPos) < pointAndClickThreshold then
-          directionIndex = directionIndex + 1
-        end
-      end
-    end
-    
-    --------------------------------------------------
-    ------------------ SMART BOI ---------------------
-    --------------------------------------------------
-    if agentType == AgentType.SmartBoiAgent then
-      shootDirection = nil
-      moveDirectionX = nil
-      moveDirectionY = nil
-      
-      -- our agent should never give up on the goal if we are forcing it there
-      local movementThreshold = pointAndClickThreshold
-      if directions and directionIndex == #directions then
-        movementThreshold = -1
-      end
-      
-      if directions and directions[directionIndex] then
-        -- print all of the grid indexes at their positions
-        printAllGridIndices(directions)
-        
-        local playerPos =  getPlayerPosition()
-        
-        local xDistToNextPos = directions[directionIndex].X - playerPos.X
-        local yDistToNextPos = directions[directionIndex].Y - playerPos.Y
-        
-        if math.abs(xDistToNextPos) > movementThreshold then
-          if xDistToNextPos > 0 then
-            moveDirectionX = ButtonAction.ACTION_RIGHT
-            shootDirection = ButtonAction.ACTION_SHOOTRIGHT
-            
-          elseif xDistToNextPos < 0 then
-            moveDirectionX = ButtonAction.ACTION_LEFT
-            shootDirection = ButtonAction.ACTION_SHOOTLEFT
-          end
-        end
-        
-        if math.abs(yDistToNextPos) > movementThreshold then
-          if yDistToNextPos < 0 then
-            moveDirectionY = ButtonAction.ACTION_UP
-            shootDirection = ButtonAction.ACTION_SHOOTUP
-            
-          elseif yDistToNextPos > 0 then
-            moveDirectionY = ButtonAction.ACTION_DOWN
-            shootDirection = ButtonAction.ACTION_SHOOTDOWN
-          end
-        end
-        
-        if directionIndex < #directions then
-          if math.abs(xDistToNextPos) < movementThreshold and math.abs(yDistToNextPos) < movementThreshold then
-            directionIndex = directionIndex + 1
-          end
-        else if goalTest() then
-            directionIndex = directionIndex + 1
-            directions = nil
-            goalTest = nil
-          end
-        end
-    end
-  end
-  printAdjacentGridIndices()
-  printNaiveDFS()
-  -- printDFS()
-  -- printAllGameEntities(getAllRoomEntities())
-end
-end
-
--- bind the MC_POST_RENDER callback to onRender
--- this event is triggered every frame, which is why we are using it to check the input
-mod:AddCallback(ModCallbacks.MC_POST_RENDER, onStep)
-
-function onUpdate()
-  updateWhereToGoNext()
-end
-mod:AddCallback(ModCallbacks.MC_POST_UPDATE, onUpdate)
-
