+++ on_damage.lua
@@ -1,0 +1,9 @@
+-------------------------------
+-------- DAMAGE EVENTS --------
+-------------------------------
+function onPlayerDamage(_,entity,_,_,source)
+  
+end
+
+-- bind the MC_ENTITY_TAKE_DMG callback for the Player to onPlayerDamage
+mod:AddCallback(ModCallbacks.MC_ENTITY_TAKE_DMG, onPlayerDamage, EntityType.ENTITY_PLAYER)
