+++ on_level_start.lua
@@ -1,0 +1,12 @@
+function setInitialLevelSearchParams()
+  visitedRooms = {}
+  roomStack = Stack:new()
+  initialRoom = {getCurrentRoom(), {}}
+  roomStack:push(initialRoom)
+end
+
+function onLevelStart()
+  setInitialLevelSearchParams()
+end
+
+mod:AddCallback(ModCallbacks.MC_POST_NEW_LEVEL, onLevelStart)
