+++ on_room_start.lua
@@ -1,0 +1,12 @@
+-- called whenever you enter a room
+function onRoomStart()
+  pointAndClickPos = nil
+  directions = nil
+  goalTest = nil
+  directionIndex = 1
+  visitedRooms[getCurrentRoom()] = true
+end
+
+-- bind the MC_POST_NEW_ROOM callback
+-- this event is triggered every time you enter a room
+mod:AddCallback(ModCallbacks.MC_POST_NEW_ROOM, onRoomStart)
